TASK: Refactor particle class to use float data types and add configuration header for simulation parameters

+++ src/particle.cpp
@@ -2,11 +2,11 @@
 
 
 // Constructor
-Particle::Particle(double d_x, double d_y, double d_vx, double d_vy, double d_ax, double d_ay, int id)
+Particle::Particle(float d_x, float d_y, float d_vx, float d_vy, float d_ax, float d_ay, int id)
     : pos(d_x, d_y), vel(d_vx, d_vy), acc(d_ax, d_ay), id_(id) {}
 
 // Update the particle's position based on its velocity and acceleration
-void Particle::Update(double d_time_step) {
+void Particle::Update(float d_time_step) {
     // Update velocity based on acceleration
     vel.x() += acc.x() * d_time_step;
     vel.y() += acc.y() * d_time_step;
